Corrected units for acceleration for IMU_v2
According to
http://www.tinkerforge.com/de/doc/Software/Bricks/IMUV2_Brick_C.html#imu-v2-brick-c-api
, the units for linear acceleration from the IMU v2 are in (1/100) m/s^2,
and the units for angular velocity are in (1/16) rad/s. Currently,
the linear acceleration raw values are multiplied by 100 instead of divided,
and the angular velocity raw values are multiplied by 16 instead of divided.
This causes abnormally high values to be reported to ROS.

After this commit,
the tinkerforge_sensors_core will correctly convert those raw values to
m/s^2 and rad/s as expected in a `sensor_msgs/Imu` message.

Secondly, the current version swaps the order of x, y, z and w quaternion values
for both IMU v1 and v2. This is not required for the IMU v2. After this commit,
the IMU v2 quaternion raw values will be directly mapped to the `sensor_msgs/Imu` output fields.